Add fade in animation to project cards
